Node-preprint divorce code review cleanup. - Remove ordering from PreprintFilesList - Move get_default_send_email_type and get_related_resource off of the node/preprint contributor serializers and add to the serializer_context on the view instead. - No need to define get_serializer_class on PreprintFilesList - Only annotate nodes queryset with "preprints_exist" if user is attempting to filter on preprints - Add a preprint manager to consolidate queries for whether a user can view particular preprints - Delete new permission groups created from PreprintService -> Preprint model rename, and then rename old groups with new group names - Remove root_folder off of Preprint model; make a property instead since this information is redundant.

+++ scripts/remove_after_use/node_preprint_es.py
@@ -9,7 +9,7 @@ from website.search.elastic_search import delete_doc
 from osf.models import Preprint, AbstractNode
 logger = logging.getLogger(__name__)
 
-# To run: docker-compose run --rm web python -m scripts.remove_after_use.node_preprint_search_mig
+# To run: docker-compose run --rm web python -m scripts.remove_after_use.node_preprint_es
 def main(dry=True):
     """
     Temporary script for updating elastic search after the node-preprint divorce
@@ -17,13 +17,19 @@ def main(dry=True):
     - Adds these nodes to the index, this time categorized as nodes
     - Adds preprints to the index, categorized as preprints
     """
-    with transaction.atomic():
-        for p in Preprint.objects.all():
+    for p in Preprint.objects.all():
+        logger.info('Adding preprint {} to index.'.format(p._id))
+        if not dry:
             p.update_search() # create new index for preprint
-            delete_doc(p.node._id, p.node, category='preprint') # delete old index for node categorized as a preprint
-            p.node.update_search() # create new index for node (this time categorized as a node)
-        if dry:
-            raise Exception('Abort Transaction - Dry Run')
+        if p.node:
+            logger.info('Deleting node {} from index, with category preprint'.format(p.node._id))
+            if not dry:
+                delete_doc(p.node._id, p.node, category='preprint') # delete old index for node categorized as a preprint
+            logger.info('Creating new index for node {}, with category node.'.format(p.node._id))
+            if not dry:
+                p.node.update_search() # create new index for node (this time categorized as a node)
+    if dry:
+        raise Exception('Abort Transaction - Dry Run')
 
 
 if __name__ == '__main__':
